gobject: CCClass use float maxDist for bbox cull (83.79->85.14%)
PSVECMag returns float; keep maxDist single-precision so the bounding-box
subtractions emit fsubs (matching target) instead of double fsub. Cast to
double only for the angle math.

+++ src/gobject.cpp
@@ -2185,13 +2185,13 @@ CGObject* CGObject::CCClass(int useBodyRadius, int classMask, float yOffset, Vec
     origin.z = m_worldPosition.z;
 
     PSVECSubtract(targetPos, &origin, &toTarget);
-    const double maxDist = static_cast<double>(PSVECMag(&toTarget));
-    if (static_cast<double>(sZeroFloat) == maxDist) {
+    const float maxDist = PSVECMag(&toTarget);
+    if (static_cast<double>(sZeroFloat) == static_cast<double>(maxDist)) {
         return 0;
     }
 
     PSVECNormalize(&toTarget, &targetDir);
-    const double maxAngle = static_cast<double>(static_cast<float>(atan2(static_cast<double>(radius), maxDist)));
+    const double maxAngle = static_cast<double>(static_cast<float>(atan2(static_cast<double>(radius), static_cast<double>(maxDist))));
     float bestDist = sLargeDistance;
     best = 0;
